Moved validation-models to models-bundle
