+++ address/address-validator/src/main/java/internship/validators/addressValidator/AddressValidator.java
@@ -3,13 +3,17 @@ package internship.validators.addressValidator;
 import internship.dao.addressDAO.AddressDAO;
 import internship.dao.userDAO.UserDAO;
 import internship.models.addressModel.Address;
-import internship.validators.addressValidator.response.ValidationError;
-import internship.validators.addressValidator.response.ValidationResult;
+import internship.models.addressModel.ValidationError;
+import internship.models.addressModel.ValidationResult;
 
 import java.util.List;
 import java.util.Set;
 
 public class AddressValidator implements IAddressValidator {
+
+    //=================================
+    // Smth for services connection
+    //=================================
 
     private UserDAO userDao;
     void setUserDao(UserDAO userDao) {
@@ -40,7 +44,7 @@ public class AddressValidator implements IAddressValidator {
         return vr;
     }
 
-    //TODO: Раньше валидация удаляла некорректные адреса из списка. Теперь не удаляет.
+    //TODO: Раньше валидация удаляла некорректные адреса из списка. Теперь не удаляет. Она и не должна, но другой метод мог сломаться.
     public ValidationResult validate(List<Address> addresses){
         ValidationResult vr = new ValidationResult();
         if (addresses==null){
@@ -71,7 +75,8 @@ public class AddressValidator implements IAddressValidator {
 
     /**
      * Проверить существует ли пользователь в базе данных
-     * @param passport
+     * @param vr Куда добавлять сообщения об ошибках
+     * @param passport Номер паспорта для проверки
      */
     private void validateUserForExistance(ValidationResult vr, Long passport) {
         if (userDao.findUserByPassport(passport) == null) {
@@ -79,6 +84,11 @@ public class AddressValidator implements IAddressValidator {
         }
     }
 
+    /**
+     * Провалидировать содержимое тега <Users></Users>
+     * @param vr Куда добавлять сообщения об ошибках
+     * @param users Список пользователей для проверки
+     */
     private void validateUsers(ValidationResult vr, Set<Long> users){
         if (users==null || users.size()==0){
             vr.addError(new ValidationError("Users", "Адрес должен иметь как минимум одного зарегестрированного в нём пользователя"));
@@ -90,30 +100,66 @@ public class AddressValidator implements IAddressValidator {
         }
     }
 
+    /**
+     * Провалидировать содержимое тега <country></country>
+     * @param vr Куда добавлять сообщения об ошибках
+     * @param country Значение для проверки
+     */
     private void validateCountry(ValidationResult vr, String country) {
         validateCommonLocation(vr, "Country", country);
     }
 
+    /**
+     * Провалидировать содержимое тега <region></region>
+     * @param vr Куда добавлять сообщения об ошибках
+     * @param region Значение для проверки
+     */
     private void validateRegion(ValidationResult vr, String region) {
         validateCommonLocation(vr, "Region", region);
     }
 
+    /**
+     * Провалидировать содержимое тега <city></city>
+     * @param vr Куда добавлять сообщения об ошибках
+     * @param city Значение для проверки
+     */
     private void validateCity(ValidationResult vr, String city) {
         validateCommonLocation(vr, "City", city);
     }
 
+    /**
+     * Провалидировать содержимое тега <street></street>
+     * @param vr Куда добавлять сообщения об ошибках
+     * @param street Значение для проверки
+     */
     private void validateStreet(ValidationResult vr, String street) {
         validateCommonLocation(vr, "Street", street);
     }
 
+    /**
+     * Провалидировать содержимое тега <houseNumber></houseNumber>
+     * @param vr Куда добавлять сообщения об ошибках
+     * @param houseNumber Значение для проверки
+     */
     private void validateHouseNumber(ValidationResult vr, String houseNumber){
         validateCommonNumber(vr, "houseNumber", houseNumber);
     }
 
+    /**
+     * Провалидировать содержимое тега <apartmentNumber></apartmentNumber>
+     * @param vr Куда добавлять сообщения об ошибках
+     * @param apartmentNumber Значение для проверки
+     */
     private void validateApartmentNumber(ValidationResult vr, String apartmentNumber){
         validateCommonNumber(vr, "apartmentNumber", apartmentNumber);
     }
 
+    /**
+     * Провалидировать тег Country\Region\City\Street. Потому что к ним предъявляются одинаковые требования.
+     * @param vr Куда добавлять сообщения об ошибках
+     * @param cause Название валидируемого тега
+     * @param value Значение для проверки
+     */
     private void validateCommonLocation(ValidationResult vr, String cause, String value){
         if (value == null) {
             vr.addError(new ValidationError(cause,"Значение не может быть пустым"));
@@ -132,6 +178,12 @@ public class AddressValidator implements IAddressValidator {
         }
     }
 
+    /**
+     * Провалидировать тег houseNumber\apartmentNumber. Потому что к ним предъявляются одинаковые требования.
+     * @param vr Куда добавлять сообщения об ошибках
+     * @param cause Название валидируемого тега
+     * @param value Значение для проверки
+     */
     private void validateCommonNumber(ValidationResult vr, String cause, String value){
         if (value==null || value.length()==0){
             vr.addError(new ValidationError(cause, "Значение не может быть пустым"));
