+++ address/address-validator/src/main/java/internship/validators/addressValidator/IAddressValidator.java
@@ -1,22 +1,27 @@
 package internship.validators.addressValidator;
 
 import internship.models.addressModel.Address;
-import internship.validators.addressValidator.response.ValidationResult;
+import internship.models.addressModel.ValidationResult;
 
 import java.util.List;
 
 public interface IAddressValidator {
 
     /**
-     * Проверяет адрес на корректность заполненных данных.
+     * Проверяет адрес на корректность заполнения данных
+     * @param address Адрес для проверки
+     * @return Возвращает результат проверки
      */
     ValidationResult validate(Address address);
 
     /**
      * Проверяет список адресов на корректность.
+     * @param addresses Адреса для проверки
+     * @return Возвращает результат проверки
      */
     ValidationResult validate(List<Address> addresses);
 
+    // TODO: Что это? По логике этого здесь не должно быть.
     boolean isUserDAOUp();
 
 }
